Pointless commit to leave note
We need to keep track of packets until they get ACKed and add a timer
and stuff. Would it makes sense to do a dictionary of structs? Have
they key be the sequence num and store the data and timer, then delete
when ACKed? Or is there an easier way?